dto interpreter

+++ app/src/main/java/com/example/weatherapp/data/repository/WeatherRepositoryImpl.kt
@@ -1,15 +1,11 @@
 package com.example.weatherapp.data.repository
 
 import android.content.SharedPreferences
-import android.provider.Settings.Global.putLong
-import android.util.Log
 import androidx.core.content.edit
 import com.example.weatherapp.common.DataState
 import com.example.weatherapp.data.local.LocalDataSource
-import com.example.weatherapp.data.local.dao.TemperatureDao
-import com.example.weatherapp.data.local.entity.WeatherEntity
 import com.example.weatherapp.data.remote.RemoteDataSource
-import com.example.weatherapp.data.remote.dto.WeatherDto
+import com.example.weatherapp.data.utils.MyEntityToDomainMapper
 import com.example.weatherapp.domain.model.Weather
 import com.example.weatherapp.domain.repository.WeatherRepository
 import java.util.concurrent.TimeUnit
@@ -18,30 +14,11 @@ import javax.inject.Inject
 class WeatherRepositoryImpl @Inject constructor(
     private val remoteDataSource: RemoteDataSource,
     private val localDataSource: LocalDataSource,
-    private val sharedPreferences: SharedPreferences
+    private val sharedPreferences: SharedPreferences,
+    private val mapper:MyEntityToDomainMapper
 ) :
     WeatherRepository {
 
-    private suspend fun getWeatherRemote(latitude: Double, longitude: Double): Weather {
-        return when (val weatherDataState =
-            remoteDataSource.getWeatherRemote(latitude, longitude)) {
-            is DataState.Loading -> {
-                // În cazul în care este încărcare, poți arunca o excepție sau să gestionezi altfel această stare
-                throw IllegalStateException("Eroare: Încărcare în curs")
-            }
-            is DataState.Success -> {
-                val weather = weatherDataState.data.toWeather()
-                weather // Returnezi obiectul Weather din starea de succes
-            }
-            is DataState.Error -> {
-                // În cazul în care este eroare, poți arunca o excepție sau să gestionezi altfel această stare
-                throw IllegalStateException("Eroare la preluarea datelor meteorologice: ${weatherDataState}")
-            }
-            else -> {
-                throw IllegalStateException("Eroare la preluarea datelor meteorologice: ${weatherDataState}")
-            }
-        }
-    }
 
     override suspend fun getWeather(latitude: Double, longitude: Double): Weather {
         val lastApiCallTime = sharedPreferences.getLong(KEY_LAST_API_CALL_TIME, 0)
@@ -49,19 +26,21 @@ class WeatherRepositoryImpl @Inject constructor(
         val differenceInMinutes = TimeUnit.MILLISECONDS.toMinutes(currentTime - lastApiCallTime)
         return if (differenceInMinutes >= 1) {
             // Se face un apel la API pentru datele actuale
-            val weather = getWeatherRemote(latitude, longitude)
-            localDataSource.deleteAll()
-            saveWeatherToLocalDatabase(weather)
-            updateLastApiCallTime(currentTime)
-            weather
+            val weather = remoteDataSource.getWeatherRemote(latitude, longitude)
+            if(weather is DataState.Success) {
+                localDataSource.deleteAll()
+                saveWeatherToLocalDatabase(weather.data)
+                updateLastApiCallTime(currentTime)
+                weather.data
+            }
+            mapper.mapToDomain(localDataSource.getWeatherLocal())
         } else {
-            // Se preiau datele din baza de date locală
-            localDataSource.getWeatherLocal().toWeather()
+            mapper.mapToDomain(localDataSource.getWeatherLocal())
         }
     }
     private suspend fun saveWeatherToLocalDatabase(weather: Weather) {
         // Salvează datele în baza de date locală
-        localDataSource.insertWeather(weather.toWeatherEntity())
+        localDataSource.insertWeather(mapper.mapToSource(weather))
     }
 
     private fun updateLastApiCallTime(currentTime: Long) {
@@ -71,28 +50,6 @@ class WeatherRepositoryImpl @Inject constructor(
         }
     }
 
-    private fun WeatherDto.toWeather(): Weather {
-        return Weather(
-            temperature = this.currentTemperatureDto.temperature
-        )
-    }
-
-    private fun WeatherEntity.toWeather(): Weather{
-        return Weather(
-            temperature = temperature
-        )
-    }
-
-    private fun WeatherDto.toWeatherEntity(): WeatherEntity {
-        return WeatherEntity(
-            temperature = this.currentTemperatureDto.temperature
-        )
-    }
-    private fun Weather.toWeatherEntity():WeatherEntity{
-        return WeatherEntity(
-            temperature = temperature
-        )
-    }
     companion object {
         const val KEY_LAST_API_CALL_TIME = "LAST_API_CALL_TIME"
     }
